lst/requirements.txt
black==25.1.0
blinker==1.9.0
click==8.1.8
confluent-kafka==2.8.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mypy-extensions==1.0.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pymssql==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==1.4.46
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
